*** = high important **  = medium *   = low _   = not
(**) If the focus changes then it seems [seems is qualitative to why not what] to allow typing back in the terminal rather than stopping and forcing into
the Popover

(***) Unable to Do for example:
Ctrl + Shift + Plus [+] (i.e. Ctrl + Plus [+])
It gets put as above and
Ctrl + [+] the Shift [=] is being recorded as a Shift [=] rather than the [+] which is what the shift would result in

(**) Can have duplicate settings i.e. Ctrl + Shift + V which is normally paste can also be present somewhere else like find.

(**) Can't exit out if in modal mode.
Also can't move but can resize just super unintuitive
- Should switch out of Popup and see if we can get another container (new custom tab?)

(**) Pressing Esc closes out of both the popover and the keybinds page.

(_) Cursor Click links functionality likely want to make a setting inside of called "orchestrated key" or something maybe make it more cosmic/popos name for it but a key for just general settings...? That seems kind of dumb it should just be like Open Link [Mod] it takes a single mod and it should know that it takes mouse button or something... This needs more logic but likely we have a toggle button in settings which says Open Link [Mouse Button] -> [Mod] + [Mouse Button].

(_) I think we should generalize our calls to things as Mod + Keybind [Modifier(s)] so we can move away from calling it Chords or whatever - not sure.